move non-components to helpers and remove useless call
